flatpakifier.bash: prepare for native Mumble

+++ bash/flatpakifier.bash
@@ -32,7 +32,9 @@ ln -sfv $FLATPAKEXPORTS/org.gajim.Gajim $TARGETDIR/gajim
 if [ ! -f /usr/bin/okular ]; then
 	ln -sfv $FLATPAKEXPORTS/org.kde.okular $TARGETDIR/okular
 fi
-ln -sfv $FLATPAKEXPORTS/info.mumble.Mumble $TARGETDIR/mumble
+if [ ! -f /usr/bin/mumble ]; then
+	ln -sfv $FLATPAKEXPORTS/info.mumble.Mumble $TARGETDIR/mumble
+fi
 
 # Electron apps that actually need wrappers and flags
 #ln -sfv $FLATPAKEXPORTS/com.visualstudio.code $TARGETDIR/code
